Create a listener for peers using the signalling channel
