+++ src/constants.ts
@@ -3,6 +3,8 @@
 export const CODE_P2P = 421
 export const CODE_CIRCUIT = 290
 
+export const P2P_WEBRTC_STAR_ID = 'p2p-webrtc-star'
+
 // Time to wait for a connection to close gracefully before destroying it
 // manually
 export const CLOSE_TIMEOUT = 2000
