+++ src/http-server.browser.ts
@@ -1,0 +1,5 @@
+export const http = {
+  createServer() {
+    throw new Error('HTTP Server can not be created in the browser!')
+  }
+};
