+++ src/http-server.ts
@@ -1,0 +1,3 @@
+import http from 'http'
+
+export { http };
