--- src/listener.browser.ts
@@ -1,4 +1,0 @@
-
-export function createListener () {
-  throw new Error('WebRTCDirect Servers can not be created in the browser!')
-}
